+++ managed/CounterStrikeSharp.API/Core/Hosting/IScriptHostConfiguration.cs
@@ -7,25 +7,25 @@
 public interface IScriptHostConfiguration
 {
     /// <summary>
-    /// Gets the absolute path to the directory that contains all CounterStrikeSharp files.
+    /// Gets the absolute path to the directory that contains CounterStrikeSharp files.
     /// e.g. /game/csgo/addons/counterstrikesharp
     /// </summary>
     string RootPath { get; }
     
     /// <summary>
-    /// Gets the absolute path to the directory that contains all CounterStrikeSharp plugins.
+    /// Gets the absolute path to the directory that contains ounterStrikeSharp plugins.
     /// e.g. /game/csgo/addons/counterstrikesharp/plugins
     /// </summary>
     string PluginPath { get; }
     
     /// <summary>
-    /// Gets the absolute path to the directory that contains all CounterStrikeSharp configs.
+    /// Gets the absolute path to the directory that contains CounterStrikeSharp configs.
     /// e.g. /game/csgo/addons/counterstrikesharp/configs
     /// </summary>
     string ConfigsPath { get; }
     
     /// <summary>
-    /// Gets the absolute path to the directory that contains all CounterStrikeSharp game data.
+    /// Gets the absolute path to the directory that contains CounterStrikeSharp game data.
     /// e.g. /game/csgo/addons/counterstrikesharp/gamedata
     /// </summary>
     string GameDataPath { get; }
