+++ managed/CounterStrikeSharp.API/Modules/Utils/ChatColors.cs
@@ -14,29 +14,29 @@
  *  along with CounterStrikeSharp.  If not, see <https://www.gnu.org/licenses/>. *
  */
 
-namespace CounterStrikeSharp.API.Modules.Utils
+namespace CounterStrikeSharp.API.Modules.Utils;
+
+public class ChatColors
 {
-    public class ChatColors
-    {
-        public static char Default = '\x01';
-        public static char White = '\x01';
-        public static char Darkred = '\x02';
-        public static char Green = '\x04';
-        public static char LightYellow = '\x09';
-        public static char LightBlue = '\x0B';
-        public static char Olive = '\x05';
-        public static char Lime = '\x06';
-        public static char Red = '\x07';
-        public static char LightPurple = '\x03';
-        public static char Purple = '\x0E';
-        public static char Grey = '\x08';
-        public static char Yellow = '\x09';
-        public static char Gold = '\x10';
-        public static char Silver = '\x0A';
-        public static char Blue = '\x0B';
-        public static char DarkBlue = '\x0C';
-        public static char BlueGrey = '\x0A';
-        public static char Magenta = '\x0E';
-        public static char LightRed = '\x0F';
-}
+    public static char Default = '\x01';
+    public static char White = '\x01';
+    public static char Darkred = '\x02';
+    public static char Green = '\x04';
+    public static char LightYellow = '\x09';
+    public static char LightBlue = '\x0B';
+    public static char Olive = '\x05';
+    public static char Lime = '\x06';
+    public static char Red = '\x07';
+    public static char LightPurple = '\x03';
+    public static char Purple = '\x0E';
+    public static char Grey = '\x08';
+    public static char Yellow = '\x09';
+    public static char Gold = '\x10';
+    public static char Silver = '\x0A';
+    public static char Blue = '\x0B';
+    public static char DarkBlue = '\x0C';
+    public static char BlueGrey = '\x0A';
+    public static char Magenta = '\x0E';
+    public static char LightRed = '\x0F';
+    public static char Orange = '\x10';
 }
